python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai>=1.84.0
PyMuPDF==1.23.3
nltk==3.8.1
tiktoken==0.5.1
Pillow>=10.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
Pillow==10.0.1
python-dotenv==1.0.0
pymongo==4.6.0
pinecone-client==2.2.4